requirements.txt
flask
flask-cors
requests
beautifulsoup4
gunicorn
Flask
Flask-SQLAlchemy
Flask-Migrate
psycopg2-binary # The driver to connect to PostgreSQL
cloudinary      # The library for uploading images
python-dotenv   # To manage secret keys locally